+++ xmhf/src/xmhf-core/xmhf-runtime/xmhf-xcphandler/arch/x86_64/xcph-x86_64.c
@@ -151,7 +151,7 @@ u8 * xmhf_xcphandler_arch_get_idt_start(void){
 
 
 //EMHF exception handler hub
-void xmhf_xcphandler_arch_hub(u32 vector, struct regs *r){
+void xmhf_xcphandler_arch_hub(uintptr_t vector, struct regs *r){
 #ifdef __XMHF_X86_64__
     (void) vector;
     (void) r;
